feat: added nav-bar, router login, styles main-page container and header
