+++ src/pages/mainPage/ui/main-page.module.scss
@@ -1,0 +1,4 @@
+.mainContainer {
+    max-width: 1160px;
+    margin: 0 auto;
+}
