+++ src/pages/mainPage/ui/mainPage.tsx
@@ -1,11 +1,12 @@
 import { MainButton } from "@/shared/ui";
 import { Header } from "@/widgets/header";
 import Link from "next/link";
+import styles from "./main-page.module.scss";
 
 export const MainPage = () => {
     return (
-        <main>
-            <Header/>
+        <main className={styles.mainContainer}>
+            <Header isLoggedIn/>
 
         </main>
     )
