+++ src/widgets/nav-bar/index.ts
@@ -1,0 +1,1 @@
+export { NavBar } from "./ui/nav-bar";
