+++ src/widgets/nav-bar/ui/nav-bar.module.scss
@@ -1,0 +1,18 @@
+@import '../../../shared/style/variables.scss';
+
+.header__navList {
+    display: flex;
+    gap: 32px;
+    margin: 0;
+    padding: 0;
+    align-items: center;
+}
+
+.header__navItem {
+    list-style: none;
+    color: $--White;
+    font-family: 'Open Sans';
+    font-size: 22px;
+    font-weight: 400;
+    line-height: 32px;
+}
