--- src/widgets/navBar/index.ts
@@ -1,1 +1,0 @@
-export { NavBar } from "./ui/navBar";
