--- src/widgets/navBar/ui/navBar.tsx
@@ -1,11 +1,0 @@
-import { useHover } from "@/shared/hooks";
-
-import { NotificationItem } from "@/entities/notification";
-import { LoginForm } from "@/features/authentication";
-import cls from './navBar.module.scss'
-
-export const NavBar = () => {
-    return (
-        <div></div>
-    )
-    }
